+++ src/cmd_rgr/field.rs
@@ -9,13 +9,15 @@ pub fn make_subcommand<'a>() -> Command<'a> {
         .about("Create/append ranges from fields")
         .after_help(
             r###"
-Field index starting from 1
+Field numbers start with 1.
 
 Example:
 
     rgr field tests/Atha/chr.sizes --chr 1 --start 2 -a -s
 
     rgr field tests/spanr/NC_007942.gff -H --chr 1 --start 4 --end 5 --strand 7 --eq 3:tRNA --ne '7:+'
+
+    rgr field tests/rgr/ctg.tsv --chr 2 --start 3 --end 4 -H -f 6,1
 
 "###,
         )
@@ -45,7 +47,14 @@ Example:
                 .long("sharp")
                 .short('s')
                 .takes_value(false)
-                .help("Write the lines starting with a # without changes. The default is to ignore them"),
+                .help("Write the lines starting with a `#` without changes. The default is to ignore them"),
+        )
+        .arg(
+            Arg::new("fields")
+                .long("fields")
+                .short('f')
+                .takes_value(true)
+                .help("Writes selected fields and the generated range field, in the order listed"),
         )
         .arg(
             Arg::new("chr")
@@ -105,7 +114,7 @@ pub fn execute(args: &ArgMatches) -> std::result::Result<(), Box<dyn std::error:
     //----------------------------
     // Options
     //----------------------------
-    let is_append = args.is_present("append");
+    let mut is_append = args.is_present("append");
     let is_header = args.is_present("header");
     let is_sharp = args.is_present("sharp");
 
@@ -132,6 +141,21 @@ pub fn execute(args: &ArgMatches) -> std::result::Result<(), Box<dyn std::error:
         })
     } else {
         0
+    };
+
+    let fields: Vec<usize> = if args.is_present("fields") {
+        is_append = true;
+
+        let mut ints: Vec<i32> = vec![];
+        let parts: Vec<&str> = args.value_of("fields").unwrap().split(',').collect();
+        for p in parts {
+            let intspan = IntSpan::from(p);
+            intspan.elements().iter().for_each(|e| ints.push(*e));
+        }
+
+        ints.iter().map(|e| *e as usize).collect()
+    } else {
+        vec![]
     };
 
     let mut eq_of: BTreeMap<usize, String> = BTreeMap::new();
@@ -171,10 +195,26 @@ pub fn execute(args: &ArgMatches) -> std::result::Result<(), Box<dyn std::error:
     for infile in args.values_of("infiles").unwrap() {
         let reader = reader(infile);
         'line: for (i, line) in reader.lines().filter_map(|r| r.ok()).enumerate() {
+            let parts: Vec<&str> = line.split('\t').collect();
+
+            // the header line
             if is_header {
                 if i == 0 {
                     if is_append {
-                        writer.write_fmt(format_args!("{}\t{}\n", line, "range"))?;
+                        if fields.is_empty() {
+                            writer.write_fmt(format_args!("{}\t{}\n", line, "range"))?;
+                        } else {
+                            let selected: Vec<String> = fields
+                                .iter()
+                                .map(|e| parts.get(*e - 1).unwrap().to_string())
+                                .collect();
+
+                            writer.write_fmt(format_args!(
+                                "{}\t{}\n",
+                                selected.join("\t"),
+                                "range"
+                            ))?;
+                        }
                     } else {
                         writer.write_fmt(format_args!("{}\n", "range"))?;
                     }
@@ -189,12 +229,10 @@ pub fn execute(args: &ArgMatches) -> std::result::Result<(), Box<dyn std::error:
                 continue 'line;
             }
 
-            let fields: Vec<&str> = line.split('\t').collect();
-
             // --eq and --ne
             if !eq_of.is_empty() {
                 for (k, v) in &eq_of {
-                    let val = fields.get(k - 1).unwrap();
+                    let val = parts.get(k - 1).unwrap();
                     if val.to_string() != *v {
                         continue 'line;
                     }
@@ -202,7 +240,7 @@ pub fn execute(args: &ArgMatches) -> std::result::Result<(), Box<dyn std::error:
             }
             if !ne_of.is_empty() {
                 for (k, v) in &ne_of {
-                    let val = fields.get(k - 1).unwrap();
+                    let val = parts.get(k - 1).unwrap();
                     if val.to_string() == *v {
                         continue 'line;
                     }
@@ -210,17 +248,17 @@ pub fn execute(args: &ArgMatches) -> std::result::Result<(), Box<dyn std::error:
             }
 
             // build ranges
-            let chr = fields.get(idx_chr - 1).unwrap();
+            let chr = parts.get(idx_chr - 1).unwrap();
             let strand = if idx_strand == 0 {
                 ""
             } else {
-                fields.get(idx_strand - 1).unwrap()
+                parts.get(idx_strand - 1).unwrap()
             };
-            let start = fields.get(idx_start - 1).unwrap().parse::<i32>().unwrap();
+            let start = parts.get(idx_start - 1).unwrap().parse::<i32>().unwrap();
             let end = if idx_end == 0 {
                 start
             } else {
-                fields.get(idx_end - 1).unwrap().parse::<i32>().unwrap()
+                parts.get(idx_end - 1).unwrap().parse::<i32>().unwrap()
             };
 
             let rg = Range {
@@ -236,7 +274,16 @@ pub fn execute(args: &ArgMatches) -> std::result::Result<(), Box<dyn std::error:
             //----------------------------
             let new_line: String;
             if is_append {
-                new_line = format!("{}\t{}", fields.join("\t"), rg);
+                if fields.is_empty() {
+                    new_line = format!("{}\t{}", parts.join("\t"), rg);
+                } else {
+                    let selected: Vec<String> = fields
+                        .iter()
+                        .map(|e| parts.get(*e - 1).unwrap().to_string())
+                        .collect();
+
+                    new_line = format!("{}\t{}", selected.join("\t"), rg);
+                }
             } else {
                 new_line = rg.to_string();
             }