+++ src/rgr.rs
@@ -31,3 +31,4 @@ fn main() -> std::io::Result<()> {
 }
 
 // TODO: `rgr annotate`
+// TODO: `rgr sort`